airflow/requirements.txt
apache-airflow-providers-google
google-cloud-storage
google-cloud-bigquery
pandas
requests
pyarrow 